Prediction of sample arrays with the SAGA model.

saga_predict uses a Support Vector Machine (SVM) with a radial kernel to classify user samples either as transforming or untransforming. The SVM model is built with the integrated SAGA training data set and a toplist of relevant probes which are able to differentiate assay data into the mentioned risk states. The SVM is further optimized on the SAGA training data. Use at your own risk!

saga_predict(
  samplepath,
  matrix.train,
  labels.train,
  matrix.unknown,
  pData.Test,
  writeFile = 0,
  showRoc = 1
)

Arguments

samplepath
path to sample data

matrix.train
normalized, probe-averaged and batch-corrected SAGA training data.

labels.train
class labels (factors) for SAGA training data. Can either be "transforming" or "untransforming".

matrix.unknown
matrix of sample data with array names as row names and probes as column names.

pData.Test
SAGA sample or test data matrix

writeFile
default=1 for writing results to the sample folder

showRoc
default=1 for showing the ROC curve of the model performance; alternative: 0 for showing naught

Value

predictions Data frame with three columns. Column one shows the sample names, the second column shows the decision values of the svm function and column thress shows the predictions for the query assays either as transforming or untransforming.