Import user data for SAGA model building

The saga_import function loads SAGA data from textfiles in a user specified path and prepares the necessary objects for later classification of unknown arrays.

saga_import(smplpath, showjoint = 0)

Arguments

smplpath
path to the saga data folder with the user samples.

showjoint
can be 1 or 0. In case of 1 it will show a boxplot of the unnormalised data (model + user data). Default is 0.

Value

matrix.joint n-by-m matrix of joint data (SAGA data + user samples).

pData.joint joint target matrix of both SAGA data and user samples.

pData.user target matrix of user samples.

filelist list of user array names.

pData phenotype target matrix of SAGA data

SAGA_Data SAGA core data; needs to loaded from sagadata

Details

In addition to the sample files, a user-defined targets file must be provided. It contains the necessary sample names, batch information etc. You can use the saga_targets function to create a targets file for you. However, in any case, the targets file must be named "SAGA_USER_Samples.txt" and must be placed in the same folder together with the sample files.

PLEASE NOTE: The saga package works with raw Agilent data. There is no need for prior data preprocessing. The saga function will take care of this.